[Refactor] detail -> budgetdetail 이름 변경
