+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:name=".ui.budget.ProcedureDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.budget.detail.main.BudgetDetailActivity"
+            android:name=".ui.budget.budgetdetail.main.BudgetDetailActivity"
             android:exported="false" />
         <activity
             android:name=".ui.budget.BudgetContentActivity"
