+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budget/budgetdetail/main/BudgetDetailFactory.kt
@@ -1,4 +1,4 @@
-package kr.sparta.tripmate.ui.viewmodel.budget.detail.main
+package kr.sparta.tripmate.ui.viewmodel.budget.budgetdetail.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
