+++ app/src/main/res/layout/activity_budget_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.budget.detail.main.BudgetDetailActivity">
+    tools:context=".ui.budget.budgetdetail.main.BudgetDetailActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/budget_detail_toolbar"
