+++ app/src/main/res/layout/fragment_budget_detail_statistics.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.budget.detail.statistics.BudgetDetailStatisticsFragment">
+    tools:context=".ui.budget.budgetdetail.statistics.BudgetDetailStatisticsFragment">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/budget_statistics_scrollview"
